Refactor into standalone file
